Refactor Nagios XI configurations and scripts.
Removed obsolete run configurations and unnecessary restarts in scripts. Updated MySQL permissions to support broader access, replaced hardcoded service commands with loops, and adjusted scripting for Docker compatibility. All changes aim to improve maintainability and deployment flexibility.

+++ nagios_xi_src/nagiosxi/nagiosxi/basedir/scripts/restart_nagios_with_export.sh
@@ -45,12 +45,16 @@ echo "--------------------------------------"
 if [ $ret -eq 0 ]; then
 
     # Restart Nagios
-    sudo $BASEDIR/manage_services.sh restart nagios
-    ret=$?
-    if [ $ret -gt 0 ]; then
-        # Remove LOCKFILE
-        rm -f "$LOCKFILE"
-        exit 6
+    if [ "${DOCKER_INSTALLER}" != "True" ]; then
+     sudo $BASEDIR/manage_services.sh restart nagios
+      ret=$?
+      if [ $ret -gt 0 ]; then
+          # Remove LOCKFILE
+          rm -f "$LOCKFILE"
+          exit 6
+      fi
+    else
+      echo "Running docker installer, not restarting nagios, as it's not running."
     fi
 
     # Make a new NOM checkpoint
